Update and rename flax_text51.py to flax_text52.py

+++ examples/text_to_image/flax_text52.py
@@ -1043,7 +1043,7 @@ def main():
         state = optimizer.init(params)
         return state, params
     def get_initial_state2(params):
-        params = jax.tree_util.tree_map(lambda x: x.astype(jnp.bfloat16), params)
+        params = jax.tree_util.tree_map(lambda x: x, params)
 
         state = optimizer2.init(params)
         return state, params
